Revise Knox processing
Revised Knox-related code, particularly that used in the Knox Interactive jupyter notebook, so that it is compatible with the recent changes relating to the Run Risk Models notebook. For example, it now requires a local_epsg parameter, and the columns of the input csv can be in any order so long as their header names are specified.